CMAP: Modified Bio::GMOD::CMap::Data::Generic so that it doesn't need the cmap_object object passed in every method. Modified everything that passed the cmap_object (although technically not necessary since Generic will just discard it if passed).
CMAE:
Added a crude zoomed out view when there are too many maps to display.  It will bin a number of maps into a single glyph whose hight is determined by the number of maps in it.  It can be selected.
Modified the viewer so that only a map above a certain threshold of pixels will have it's details displayed (features and submaps).
Fixed a bug that under certain conditions made the sub-map zones move twice as fast as others.
Fixed another bug that occured when the first reference map was no longer on the screen.
Added a plug-in hook to the commit changes series.  Any plug-in at this spot can cancel the commit.
Added an example of the above plug-in.