Check aliases when importing correspondences
If a tab-delimited correspondence file lacks feature accessions, the
feature_name[12] columns will be used instead. The
Bio::GMOD::CMap::Admin::ImportCorrespondences POD implies that the
feature_name* columns may contain aliases:

    If the accession IDs are not present, both the "feature_name"
    and "aliases" are checked. For every feature matching each of the two
    feature names, a correspondence will be created.

However, aliases were not being considered. This mod incorporates a
patch I previously submitted (http://sourceforge.net/p/gmod/bugs/309/)
to update the code to be congruent with the documentation.

A regression test was added to verify the functionality. Note that the
new regression test does *not* pass when using SQLite---though oddly, it
does if it is moved to run after the original "import correspondences"
test.