Standardize column names returned by DBI
The failure of the new tab-delimited-correlation-file regression test
with DBD::SQLite was traced to the following issue. Given a query that
includes a range variable (aka "correlation name"; informally known as
"alias"), e.g.:

$rows = $dbh->selectall_arrayref(
        'select cf.feature_id from cmap_feature cf;',{Columns => {}});
print keys($rows->[0]);

DBD::MySQL and DBD::Pg print "feature_id", whereas DBD::SQLite retains
the range-variable prefix ("cf.feature_id"). The cmap code expects the
former behavior.

SQL statements in Bio::GMOD::CMap::Data::Generic that are affected by
this issue have been updated with the AS clause to explicitly get the
column names expected buy the rest of the CMap code; e.g.:

    select cf.feature_id AS feature_id from cmap_feature cf;

The new "AS" clauses are in uppercase to provide a visual clue that the
name on the right is the same as the name on the left, only without the
range variable (existing as clauses were in lowercase).

For a few SQL statements that didn't need the range variable, the range
variable was removed.

Bio::GMOD::CMap::Admin::GBrowseLiason was not updated, as it is
essentially deprecated due to supporting only the (unofficially?)
deprecated Bio::DB::GFF storage adaptor