Replace use of Algorithm::Numerical::Sample
Remove dependency Algorithm::Numerical::Sample by replacing its single use
with a variation of a concise (and more efficient, per a microbenchmark I
created) alternative that I got from a Randall Schwartz post
(http://www.perlmonks.org/?node_id=245171).

Data::Page is a prerequisite for Data::PageSet, but is not used directly
by CMap, and so probably isn't necessary to include in the list of
dependencies.