Support TrueType fonts and antialiased lines
Add support for TrueType fonts and antialiased lines for GIF and PNG
images. Because this causes an increase in the file size resulting
image files. the default image type is now GIF, which produces smaller
file sizes than PNG.

Note that TrueType fonts are now required, though if there is demand it
shouldn't be difficult to make them optional.