feature_xrefs service to generate JSON xrefs
Iliana Toneva (NCGR) contributed code to generate JSON formatted feature
xrefs via URL; e.g.:

/cmap/feature_xrefs?feature_acc=FEATURE_ACC&data_source=DATA_SOURCE

When a feature is clicked on, AJAX can be used to get the xrefs for that
feature and dynamically generate a context menu. This can be
significantly faster than the alternative of processing all generic
xrefs for each feature in the CMap configuration file area_code.