Improve map_details and feature page load times
Fix a performance regression in
Bio::GMOD::CMap::Data::Generic::get_feature_correspondence_details introduced
when the cmap_correspondence_lookup table was converted into a view by
using base tables when possible instead of the view (specifically, when
feature_id1 is supplied as an argument). For expediency, removed the
ability to use feature_id2 as an argument, though this wasn't used
anywhere in the CMap code base.